pyproject.toml
[project]
name = "birdwatcher-app-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-auth-oauthlib>=1.2.1",
    "google-cloud-secret-manager>=2.23.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.43.2",
]
